products

+++ app/Http/Controllers/Attribute/AttributeController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers\Attribute;
 
 use App\Attributes;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Attribute\StoreAttribute;
+use App\Http\Requests\Attribute\UpdateAttribute;
+use App\Http\Resources\Attribute\AttributeResource;
 use App\Traits\ApiResponser;
-use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
+use Exception;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Http\Response;
+use Illuminate\Pagination\LengthAwarePaginator;
 
 class AttributeController extends Controller
 {
@@ -16,94 +22,92 @@ class AttributeController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return AnonymousResourceCollection|Response|LengthAwarePaginator
      */
     public function index()
     {
-        $attributes = Attributes::all();
         if (request()->expectsJson() && request()->acceptsJson()){
-            return $this->showAll($attributes);
+            $attributes = Attributes::all();
+            return $this->showCollection(AttributeResource::collection($attributes));
         }
+        return null;
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     * @throws \Illuminate\Validation\ValidationException
+     * @param StoreAttribute $request
+     * @return AttributeResource|null
      */
-    public function store(Request $request)
+    public function store(StoreAttribute $request)
     {
-        $rules = [
-            'name'=>'required|min:2|max:100|unique:attributes,name',
-        ];
 
-        $this->validate($request,$rules);
 
-        $newAttribute = Attributes::create($request->only(['name']));
         if (request()->expectsJson() && request()->acceptsJson()){
-            return $this->showOne($newAttribute);
+            $newAttribute = Attributes::create($request->only(['name']));
+            return new AttributeResource($newAttribute);
         }
+
+        return null;
     }
 
     /**
      * Display the specified resource.
      *
      * @param Attributes $attribute
-     * @return \Illuminate\Http\Response
+     * @return AttributeResource|null
      */
     public function show(Attributes $attribute)
     {
         if (request()->expectsJson() && request()->acceptsJson()){
-            return $this->showOne($attribute);
+            return new AttributeResource($attribute);
         }
+
+        return null;
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param UpdateAttribute $request
      * @param Attributes $attribute
-     * @return \Illuminate\Http\Response
-     * @throws \Illuminate\Validation\ValidationException
+     * @return AttributeResource|JsonResponse|Response
      */
-    public function update(Request $request, Attributes $attribute)
+    public function update(UpdateAttribute $request, Attributes $attribute)
     {
-        $rules = [
-            'name'=>['min:2|max:100'
-                , Rule::unique($attribute->getTable())->ignore(request()->segment(3))
-            ],
+        if (request()->expectsJson() && request()->acceptsJson()){
+            $attribute->fill($request->only([
+                'name',
+            ]));
 
-        ];
+            if($attribute->isClean()){
+                return $this->errorResponse([
+                    'error'=> 'you need to specify a different value to update',
+                    'code'=> 422],
+                    422);
+            }
 
-        $this->validate($request,$rules);
-
-        $attribute->fill($request->only([
-            'name',
-        ]));
-
-        if($attribute->isClean()){
-            return $this->errorResponse([
-                'error'=> 'you need to specify a different value to update',
-                'code'=> 422],
-                422);
+            $attribute->save();
+            return new AttributeResource($attribute);
         }
-
-        $attribute->save();
-        return $this->showOne($attribute);
+        return null;
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param Attributes $attribute
-     * @return \Illuminate\Http\Response
-     * @throws \Exception
+     * @return AttributeResource|null
+     * @throws Exception
      */
     public function destroy(Attributes $attribute)
     {
-        $attribute->delete();
-        return $this->showOne($attribute);
+        if (request()->expectsJson() && request()->acceptsJson()){
+            $attribute->delete();
+            return new AttributeResource($attribute);
+        }
+
+        return null;
+
     }
 }
